feat(witnet): add features for compiling with/without node/wallet
