+++ Cargo.toml
@@ -13,6 +13,11 @@ edition = "2018"
 [workspace]
 members = ["config", "node", "crypto", "data_structures", "p2p", "storage", "wallet", "validations", "protected", "reputation", "net"]
 
+[features]
+default = ["wallet", "node"]
+wallet = ["witnet_wallet"]
+node = ["witnet_node"]
+
 [badges]
 travis-ci = { repository = "https://github.com/witnet/witnet-rust", branch = "master" }
 
@@ -26,20 +31,13 @@ log = "0.4.6"
 serde_json = "1.0.39"
 structopt = "0.2.14"
 toml = "0.4.10"
-clap = "2.33.0"
-
-witnet_config = { path = "./config" }
-witnet_node = { path = "./node" }
-witnet_crypto = { path = "./crypto" }
-witnet_data_structures = { path = "./data_structures" }
-witnet_p2p = { path = "./p2p" }
-witnet_rad = { path = "./rad" }
-witnet_storage = { path = "./storage", features = ["rocksdb-backend"] }
-witnet_util = { path = "./util" }
-witnet_validations = { path = "./validations" }
-witnet_wallet = { path = "./wallet" }
 terminal_size = "0.1.8"
 lazy_static = "1.3.0"
+
+witnet_wallet = { path = "./wallet", optional = true }
+witnet_node = { path = "./node", optional = true }
+witnet_config = { path = "./config" }
+witnet_data_structures = { path = "./data_structures" }
 
 [dependencies.serde]
 features = ["derive"]
