+++ src/cli/node/mod.rs
@@ -1,0 +1,12 @@
+#[cfg(feature = "node")]
+mod json_rpc_client;
+#[cfg(feature = "node")]
+mod with_node;
+#[cfg(not(feature = "node"))]
+mod without_node;
+
+#[cfg(feature = "node")]
+pub use with_node::*;
+
+#[cfg(not(feature = "node"))]
+pub use without_node::*;
