+++ src/cli/wallet/mod.rs
@@ -1,0 +1,10 @@
+#[cfg(feature = "wallet")]
+mod with_wallet;
+#[cfg(not(feature = "wallet"))]
+mod without_wallet;
+
+#[cfg(feature = "wallet")]
+pub use with_wallet::*;
+
+#[cfg(not(feature = "wallet"))]
+pub use without_wallet::*;
